requirements.txt
pyyaml==6.0.1
numpy==1.26.3
python-dateutil==2.8.2
matplotlib==3.8.2 